Tested Product Rest Api

+++ MusicStore/src/main/java/com/musicstore/controller/api/ProductRestController.java
@@ -1,6 +1,7 @@
 package com.musicstore.controller.api;
 
 import java.util.List;
+import java.util.Map;
 import java.util.ArrayList; 
 import java.util.Optional;
 
@@ -14,11 +15,26 @@ import org.springframework.web.server.ResponseStatusException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.musicstore.model.CartBean;
 import com.musicstore.model.ProductBean;
+import com.musicstore.model.WebUserBean;
+import com.musicstore.service.DbAdminService;
 import com.musicstore.service.DbProductService;
+import com.musicstore.service.DbProducerService;
+import com.musicstore.service.DbWebUserService;
+import com.musicstore.utility.Utility;
 
 @RestController
 public class ProductRestController {
+
+	@Autowired
+	private DbAdminService adminService; 
+
+	@Autowired
+	private DbProducerService producerService; 
+	
+	@Autowired
+	private DbWebUserService webuserService; 
 	
 	@Autowired
 	private DbProductService productService; 
@@ -40,14 +56,23 @@ public class ProductRestController {
 	}
 
 	@RequestMapping(value  ="/musicstore/api/product", method = RequestMethod.POST)
-	public ProductBean create(@RequestBody ProductBean p) {
-		return productService.create(p);
+	public ProductBean create(@RequestBody Map<String, Map<String,String>> map) {
+		ProductBean pb = Utility.productDeMap(map.get("topost"));
+		WebUserBean b = Utility.webuserDeMap(map.get("authorized"));
+		if(!adminService.isAdmin(b) && !pb.getProducer().equals(b.getMail()) || !webuserService.isWebUser(b)){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an authorized");
+		}
+		return productService.create(pb);
 	}
 	
 	@RequestMapping(value  ="/musicstore/api/product/{id}", method = RequestMethod.PUT)
-	public ProductBean update(@PathVariable int id, @RequestBody ProductBean p) {
-		
-		Optional<ProductBean> updatedProduct= productService.update(id, p);
+	public ProductBean update(@PathVariable int id, @RequestBody Map<String, Map<String,String>> map) {
+		ProductBean pb = Utility.productDeMap(map.get("topost"));
+		WebUserBean b = Utility.webuserDeMap(map.get("authorized"));
+		if(!adminService.isAdmin(b) && !pb.getProducer().equals(b.getMail())|| !webuserService.isWebUser(b)){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an authorized");
+		}
+		Optional<ProductBean> updatedProduct= productService.update(id, pb);
 		if (updatedProduct.isEmpty())
 		{
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item not found");
@@ -56,7 +81,10 @@ public class ProductRestController {
 	}
 	
 	@RequestMapping(value  ="/musicstore/api/product/{id}", method = RequestMethod.DELETE)
-	public void delete(@PathVariable int id) {
+	public void delete(@PathVariable int id, @RequestBody WebUserBean b) {
+		if(!adminService.isAdmin(b) && !getById(id).getProducer().equals(b.getMail()) || !webuserService.isWebUser(b)){
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "request by not an authorized");
+		}
 		Boolean isDeleted = productService.delete(id);
 		if (isDeleted==false)
 		{
